// examples/react/src/generated/operations.ts
export interface TCreateBossReviewInput {
  /**
   * (String)
   */
  author?: string;

  /**
   * (ID)
   */
  bossId: string;

  difficulty: TDifficulty;

  theme: TRating;
}

export interface TCreateLocationReviewInput {
  /**
   * (String)
   */
  author?: string;

  /**
   * (ID)
   */
  locationId: string;

  difficulty: TDifficulty;

  design: TRating;
}

export enum TRating {
  TERRIBLE = "TERRIBLE",

  MEH = "MEH",

  ALRIGHT = "ALRIGHT",

  AMAZING = "AMAZING",

  STELLAR = "STELLAR",
}

export enum TDifficulty {
  OKAYISH = "OKAYISH",

  HARD = "HARD",

  IMPOSSIBLE = "IMPOSSIBLE",
}